Put write-open-read integration test in separate module
