+++ haskey.cabal
@@ -110,6 +110,8 @@ test-suite haskey-properties
 test-suite haskey-integration
   main-is:             Integration.hs
   type:                exitcode-stdio-1.0
+  other-modules:
+    Integration.WriteOpenRead
 
   build-depends:
     base          >=4.7  && <5,
@@ -123,6 +125,8 @@ test-suite haskey-integration
     vector        >=0.10 && <1,
 
     QuickCheck    >=2    && <3,
+    test-framework             >=0.8  && <1,
+    test-framework-quickcheck2 >=0.3  && <1,
     haskey
 
   default-language:    Haskell2010
